notes/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==20.3.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
distlib==0.3.4
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.7.1
h11==0.13.0
idna==3.3
jellyfish==0.9.0
lib50==3.0.4
Markdown==3.3.7
markdown2==2.4.3
outcome==1.2.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.2
ptyprocess==0.7.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.4
PyYAML==5.4.1
requests==2.28.0
selenium==4.4.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
submit50==3.1.1
termcolor==1.1.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.15.1
wsproto==1.1.0
